Redefinition of functions
